+++ attributes/default.rb
@@ -34,6 +34,16 @@ default['rsyslog']['repeated_msg_reduction']    = 'on'
 default['rsyslog']['logs_to_forward']           = '*.*'
 default['rsyslog']['enable_imklog']             = true
 default['rsyslog']['config_prefix']             = '/etc'
+default['rsyslog']['default_file_template']    = nil
+default['rsyslog']['default_remote_template']   = nil
+default['rsyslog']['rate_limit_interval']       = nil
+default['rsyslog']['rate_limit_burst']          = nil
+default['rsyslog']['enable_tls']                = false
+default['rsyslog']['tls_ca_file']               = nil
+default['rsyslog']['tls_certificate_file']      = nil
+default['rsyslog']['tls_key_file']              = nil
+default['rsyslog']['tls_auth_mode']             = 'anon'
+default['rsyslog']['use_local_ipv4']            = false
 
 # The most likely platform-specific attributes
 default['rsyslog']['service_name']              = 'rsyslog'
@@ -70,12 +80,12 @@ when 'rhel'
   # format { facility => destination }
   default['rsyslog']['default_facility_logs'] = {
     '*.info;mail.none;authpriv.none;cron.none' => "#{node['rsyslog']['default_log_dir']}/messages",
-    'authpriv' => "#{node['rsyslog']['default_log_dir']}/secure",
+    'authpriv.*' => "#{node['rsyslog']['default_log_dir']}/secure",
     'mail.*' => "-#{node['rsyslog']['default_log_dir']}/maillog",
     'cron.*' => "#{node['rsyslog']['default_log_dir']}/cron",
     '*.emerg' => '*',
     'uucp,news.crit' => "#{node['rsyslog']['default_log_dir']}/spooler",
-    'local7.' => "#{node['rsyslog']['default_log_dir']}/boot.log"
+    'local7.*' => "#{node['rsyslog']['default_log_dir']}/boot.log"
   }
 else
   # format { facility => destination }