+++ metadata.rb
@@ -4,7 +4,7 @@ maintainer_email  'cookbooks@opscode.com'
 license           'Apache 2.0'
 description       'Installs and configures rsyslog'
 long_description  IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version           '1.9.1'
+version           '1.12.3'
 
 recipe            'rsyslog', 'Installs rsyslog'
 recipe            'rsyslog::client', 'Sets up a client to log to a remote rsyslog server'
@@ -86,4 +86,39 @@ attribute 'rsyslog/repeated_msg_reduction',
 attribute 'rsyslog/priv_seperation',
   :display_name => 'Privilege separation',
   :description => 'Whether or not to make use of Rsyslog privilege separation',
+  :default => 'false'
+
+attribute 'rsyslog/default_file_template',
+  :display_name => 'Default file log format template',
+  :description => 'The name of a pre-defined log format template (ie - `RSYSLOG_FileFormat`), used for local log files.'
+
+attribute 'rsyslog/default_remote_template',
+  :display_name => 'Default remote log format template',
+  :description => 'The name of a pre-defined log format template (ie - `RSYSLOG_SyslogProtocol23Format`), used for remote log forwarding.'
+
+attribute 'rsyslog/enable_tls',
+  :display_name => 'Enable TLS',
+  :description => 'Whether or not to enable TLS encryption. When enabled, forces protocol to "tcp"',
+  :default => 'false'
+
+attribute 'rsyslog/tls_ca_file',
+  :display_name => 'TLS CA file',
+  :description => 'Path to TLS CA file. Required for both server and clients.'
+
+attribute 'rsyslog/tls_certificate_file',
+  :display_name => 'TLS certificate file',
+  :description => 'Path to TLS certificate file. Required for server, optional for clients.'
+
+attribute 'rsyslog/tls_key_file',
+  :display_name => 'TLS key file',
+  :description => 'Path to TLS key file. Required for server, optional for clients.'
+
+attribute 'rsyslog/tls_auth_mode',
+  :display_name => 'TLS auth mode',
+  :description => 'Value for "$InputTCPServerStreamDriverAuthMode"/"$ActionSendStreamDriverAuthMode", determines whether client certs are validated.',
+  :default => 'anon'
+
+attribute 'rsyslog/use_local_ipv4',
+  :display_name => 'Try to use local IPv4 address',
+  :description => 'Whether or not to make use the remote local IPv4 address on cloud systems when searching for servers (where available).',
   :default => 'false'