+++ spec/spec_helper.rb
@@ -1,8 +1,4 @@
-require 'berkshelf'
 require 'chefspec'
-require 'chef/application' # CHEF-3407
+require 'chefspec/berkshelf'
 
-Berkshelf.ui.mute do
-  berksfile = Berkshelf::Berksfile.from_file('Berksfile')
-  berksfile.install(path: 'vendor/cookbooks', only: 'integration')
-end
+at_exit { ChefSpec::Coverage.report! }