+++ CHANGELOG.md
@@ -3,6 +3,44 @@ rsyslog Cookbook CHANGELOG
 This file is used to list changes made in each version of the rsyslog cookbook.
 
 
+
+
+
+v1.12.2 (2014-02-28)
+--------------------
+Fixing bug fix in rsyslog.conf
+
+
+v1.12.0 (2014-02-27)
+--------------------
+- [COOK-4021] Allow specifying default templates for local and remote
+- [COOK-4126] rsyslog cookbook fails restarts due to not using upstart
+
+
+v1.11.0 (2014-02-19)
+--------------------
+### Bug
+- **[COOK-4256](https://tickets.opscode.com/browse/COOK-4256)** - Fix syntax errors in default.conf on rhel
+
+### New Feature
+- **[COOK-4022](https://tickets.opscode.com/browse/COOK-4022)** - Add use_local_ipv4 option to allow selecting internal interface on cloud systems
+- **[COOK-4018](https://tickets.opscode.com/browse/COOK-4018)** - rsyslog TLS encryption support
+
+
+v1.10.2
+-------
+No change. Version bump for toolchain.
+
+
+v1.10.0
+-------
+### New Feature
+- **[COOK-4021](https://tickets.opscode.com/browse/COOK-4021)** - Allow specifying default templates for local and remote
+
+### Improvement
+- **[COOK-3876](https://tickets.opscode.com/browse/COOK-3876)** - Cater for setting rate limits
+
+
 v1.9.0
 ------
 ### New Feature